Debugging tests for new functionality
