+++ tests/test_parameters.py
@@ -10,7 +10,7 @@ class TestAimsControl:
     def aims_fixture_no(self) -> int:
         return int(self.ac.path.split("/")[-2])
 
-    @pytest.fixture(params=range(1, 13), autouse=True)
+    @pytest.fixture(params=range(1, 14), autouse=True)
     def aims_control(self, cwd, request, aims_calc_dir):
         self.ac = AimsControl(
             control_in=f"{cwd}/fixtures/{aims_calc_dir}/{str(request.param)}/control.in"
@@ -36,26 +36,32 @@ class TestAimsControl:
 
     @pytest.fixture
     def cube_cell_ref_files(self, cwd):
-        with open(
-            f"{cwd}/fixtures/manipulated_aims_files/cube_cell/"
-            f"{self.aims_fixture_no}/control.in",
-            "r",
-        ) as f:
-            yield f.readlines()
+        if self.aims_fixture_no != 13:
+            with open(
+                f"{cwd}/fixtures/manipulated_aims_files/cube_cell/"
+                f"{self.aims_fixture_no}/control.in",
+                "r",
+            ) as f:
+                yield f.readlines()
+
+        else:
+            yield None
 
     def test_get_keywords(self, ref_data, cwd):
         assert self.ac.get_keywords() == ref_data["keywords"][self.aims_fixture_no - 1]
 
         # Check it works for the cube files as these have multiple keywords that are the
         # same with different values
-        cube_ac = AimsControl(
-            control_in=f"{cwd}/fixtures/manipulated_aims_files/cube_cell/"
-            f"{self.aims_fixture_no}/control.in"
-        )
-        assert (
-            cube_ac.get_keywords()
-            == ref_data["cube_keywords"][self.aims_fixture_no - 1]
-        )
+        if self.aims_fixture_no != 13:
+            cube_ac = AimsControl(
+                control_in=f"{cwd}/fixtures/manipulated_aims_files/cube_cell/"
+                f"{self.aims_fixture_no}/control.in"
+            )
+            # if self.aims_fixture_no != 13:
+            assert (
+                cube_ac.get_keywords()
+                == ref_data["cube_keywords"][self.aims_fixture_no - 1]
+            )
 
     def test_get_species(self):
         cluster_species = ["O", "H"]
@@ -63,6 +69,8 @@ class TestAimsControl:
 
         if self.aims_fixture_no in [1, 2, 3, 5, 7, 9]:
             assert self.ac.get_species() == cluster_species
+        elif self.aims_fixture_no == 13:
+            assert self.ac.get_species() == ["Cu", "O", "C"]
         else:
             assert self.ac.get_species() == periodic_species
 
@@ -80,37 +88,42 @@ class TestAimsControl:
         if self.aims_fixture_no == 4:
             assert self.ac.get_default_basis_funcs() == basis_funcs[1]
 
-    def test_add_cube_cell(self, tmp_dir, cube_cell_ref_files):
-        control_path = tmp_dir / "control.in"
-        shutil.copy(self.ac.path, control_path)
-        ac = AimsControl(control_in=str(control_path))
-        try:
-            ac.add_keywords(("output", "cube total_density"))
-            ac.add_cube_cell(np.eye(3, 3) * [3, 4, 5], resolution=100)
-        except TypeError:
-            assert not ac.check_periodic()
-        else:
-            assert (
-                "".join(cube_cell_ref_files) == control_path.read_text()
-            )  # Check correct for periodic
+    def test_add_cube_cell_and_save(self, tmp_dir, cube_cell_ref_files):
+        if self.aims_fixture_no != 13:
+            control_path = tmp_dir / "control.in"
+            shutil.copy(self.ac.path, control_path)
+            ac = AimsControl(control_in=str(control_path))
+            try:
+                ac.add_keywords_and_save(("output", "cube total_density"))
+                ac.add_cube_cell_and_save(np.eye(3, 3) * [3, 4, 5], resolution=100)
+            except TypeError:
+                assert not ac.check_periodic()
+            else:
+                assert (
+                    "".join(cube_cell_ref_files) == control_path.read_text()
+                )  # Check correct for periodic
 
-    def test_add_keywords(self, tmp_dir, added_keywords_ref_files):
+    def test_add_keywords_and_save(self, tmp_dir, added_keywords_ref_files):
         control_path = tmp_dir / "control.in"
         shutil.copy(self.ac.path, control_path)
         ac = AimsControl(control_in=str(control_path))
-        ac.add_keywords(
+        ac.add_keywords_and_save(
             ("xc", "dfauto scan"),
             ("output", "cube spin_density"),
             ("output", "mulliken"),
         )
 
+        if self.aims_fixture_no == 13:
+            with open("tmp.out", "w") as f:
+                f.writelines(ac.lines)
+
         assert "".join(added_keywords_ref_files) == control_path.read_text()
 
-    def test_remove_keywords(self, tmp_dir, removed_keywords_ref_files):
+    def test_remove_keywords_and_save(self, tmp_dir, removed_keywords_ref_files):
         control_path = tmp_dir / "control.in"
         shutil.copy(self.ac.path, control_path)
         ac = AimsControl(control_in=str(control_path))
-        ac.remove_keywords("xc", "relax_geometry", "k_grid")
+        ac.remove_keywords_and_save("xc", "relax_geometry", "k_grid")
 
         assert "".join(removed_keywords_ref_files) == control_path.read_text()
